+++ tests/packages/test_dependency.py
@@ -70,6 +70,21 @@ def test_to_pep_508():
                      'or (python_version >= "3.6" and python_version < "4.0")'
 
 
+def test_to_pep_508_with_platform():
+    dependency = Dependency('Django', '^1.23')
+    dependency.python_versions = '~2.7 || ^3.6'
+
+    dependency.platform = 'linux || linux2'
+
+    result = dependency.to_pep_508()
+    assert result == (
+        'Django (>=1.23,<2.0); '
+        '((python_version >= "2.7" and python_version < "2.8") '
+        'or (python_version >= "3.6" and python_version < "4.0"))'
+        ' and (sys_platform == "linux" or sys_platform == "linux2")'
+    )
+
+
 def test_to_pep_508_wilcard():
     dependency = Dependency('Django', '*')